Add files via upload
Finally, a dashboard was created in Power BI using these queries to visualize the data in a clear and concise manner. This project aimed to provide insights into the COVID-19 pandemic and to make the data more accessible and understandable for the general public. The use of SQL and Power BI allowed for efficient data processing and visualization, making it easier to identify trends and patterns in the data. Overall, this project demonstrates the value of using data analysis tools to gain insights into important global issues like the COVID-19 pandemic.